Fix README
Steps "clone the repo" and "install package from pip" are contradictory: if you install the pip package, you don't need the repo.
However, the current release of the pip package doesn't work with the example from README sice OrpheusModel lacks **engine_kwargs in that release. So it makes more sense to install the package from the repo.